hooks化
